requirements.txt
beautifulsoup4==4.6.0
Django==3.0.5
djangotoolbox==1.8.0
future==0.16.0
jsondiff==1.1.1
nose==1.3.7
pytest==3.2.1
python-dateutil==2.7.5
requests==2.18.1
six
SQLAlchemy==1.2.19
traitlets==4.3.2
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
urllib3==1.21.1
djangorestframework==3.11.0
coreapi==2.3.3